gus_function/requirements.txt
flask
requests
google-cloud-storage
google-cloud-bigquery
